requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
async-timeout==4.0.3
attrs==23.2.0
frozenlist==1.4.1
idna==3.6
loguru==0.7.2
multidict==6.0.4
pydantic==2.5.3
pydantic_core==2.14.6
pyserial==3.5
pysolarmanv5==3.0.1
ruamel.yaml==0.18.5
ruamel.yaml.clib==0.2.8
typing_extensions==4.9.0
uModbus==1.0.4
yarl==1.9.4
